requirements.txt
flask
pypdf
python-docx
nltk
joblib
scikit-learn
numpy
reportlab
fpdf2
